Login functionality
also some refactoring and componentization of signup
