+++ src/views/index/HomeContact.js
@@ -5,11 +5,6 @@ import {
     CardHeader,
     CardBody,
     CardTitle,
-    Form,
-    Input,
-    InputGroupAddon,
-    InputGroupText,
-    InputGroup,
     Row,
     UncontrolledTooltip,
 } from "reactstrap";
